JobRouteAPI/requirements.txt
fastapi==0.100.0
uvicorn==0.27.0
pymongo==4.5.0
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
tensorflow==2.13.0
numpy==1.24.3
scikit-learn==1.2.2
pandas==1.5.3
jwt==1.3.1
seaborn==0.11.2
